+++ src/ByHttpRequestQueryLookup.php
@@ -3,8 +3,6 @@
 namespace SEQL;
 
 use MediaWiki\MediaWikiServices;
-use Onoi\Cache\CacheFactory as OnoiCacheFactory;
-use Onoi\HttpRequest\HttpRequestFactory;
 use SEQL\ByHttpRequest\JsonResponseParser;
 use SEQL\ByHttpRequest\QueryResultFetcher;
 use SMW\CacheFactory;
@@ -68,8 +66,11 @@ class ByHttpRequestQueryLookup extends SQLStore {
 	}
 
 	protected function fetchQueryResultFor( Query $query, $interwiki, $credentials = false ) {
+		$services = MediaWikiServices::getInstance();
+
 		$queryResultFetcher = new QueryResultFetcher(
-			new HttpRequestFactory( $this->newHttpResponseCache( $GLOBALS['seqlgHttpResponseCacheType'] ) ),
+			$services->getHttpRequestFactory(),
+			$services->getObjectCacheFactory()->getInstance( $GLOBALS['seqlgHttpResponseCacheType'] ),
 			$this->queryResultFactory,
 			new JsonResponseParser( new DataValueDeserializer( $query->getQuerySource() ) ),
 			$credentials
@@ -90,21 +91,6 @@ class ByHttpRequestQueryLookup extends SQLStore {
 		}
 
 		return $this->cacheFactory;
-	}
-
-	/**
-	 * Builds the Onoi cache used by the HTTP request layer. Replaces the
-	 * removed SMW `CacheFactory::newMediaWikiCompositeCache` by composing the
-	 * equivalent Onoi caches directly on top of a MediaWiki BagOStuff.
-	 */
-	private function newHttpResponseCache( $cacheType ) {
-		$objectCacheFactory = MediaWikiServices::getInstance()->getObjectCacheFactory();
-		$cacheFactory = OnoiCacheFactory::getInstance();
-
-		return $cacheFactory->newCompositeCache( [
-			$cacheFactory->newFixedInMemoryCache( 500 ),
-			$cacheFactory->newMediaWikiCache( $objectCacheFactory->getInstance( $cacheType ) )
-		] );
 	}
 
 }