+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@ This file contains the RELEASE-NOTES of the Semantic External Query Lookup (a.k.
 
 - Added support for Semantic MediaWiki 7.0
 - Raised the minimum requirements to MediaWiki 1.43 and Semantic MediaWiki 7.0
+- Replaced the `mediawiki/http-request` dependency with MediaWiki core's `HttpRequestFactory`
 
 ### 1.0.0 (2016-??-??)